\name{sisglmlss-package}
\alias{sisglmlss-package}
\docType{package}
\title{
\packageTitle{sisglmlss}
}
\description{
\packageDescription{sisglmlss}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sisglmlss}
\packageIndices{sisglmlss}
}
\author{
\packageAuthor{sisglmlss}

Maintainer: \packageMaintainer{sisglmlss}
}
\references{

Diego Franco Saldana and Yang Feng (2018) SIS: An R package for Sure Independence Screening in
Ultrahigh Dimensional Statistical Models, \emph{Journal of Statistical Software}, \bold{83}, 2, 1-25.

Jianqing Fan and Rui Song (2010) Sure Independence Screening in Generalized
Linear Models with NP-Dimensionality.  \emph{The Annals of Statistics},
\bold{38}, 3567-3604.

Jianqing Fan, Richard Samworth, and Yichao Wu (2009) Ultrahigh Dimensional
Feature Selection: Beyond the Linear Model. \emph{Journal of Machine
Learning Research}, \bold{10}, 2013-2038.

Rigby, R. A. and  Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), 
\emph{Appl. Statist.}, \bold{54}, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R.
\emph{Journal of Statistical Software}, Vol. \bold{23}, Issue 7, Dec 2007, \url{https://www.jstatsoft.org/v23/i07/}.

}

\keyword{ package }
\keyword{ distribution }
\keyword{ regression }

\seealso{
\code{\link[gamlss]{gamlss}}, \code{\link[gamlss.dist]{gamlss.dist}}, \code{\link[gamlss.lasso2]{gnet}} 
}

